Separated out Tab (previously Page) from AdministratorWindow.
