+++ modules/gui/administratorWindow.py
@@ -1,13 +1,7 @@
 from tkinter import *
 from tkinter import ttk
-from tkinter import messagebox
 import tkinter
-import PIL.Image
-import PIL.ImageTk
-import cv2
-import os
-from fpdf import FPDF
-import subprocess
+from tab import Tab
 
 # Define the elements to be laid out on each tab
 _gui_data = {
@@ -274,7 +268,7 @@ class AdministratorWindow(Tk):
 
         for i in range(len(list(body.keys()))):
             page_configuration = body[list(body.keys())[i]]
-            tab = Page(self.notebook, self, cap, 
+            tab = Tab(self.notebook, self, cap, 
                     on_ear_change, initial_ear, 
                     on_low_contrast, initial_low_contrast,
                     on_high_contrast, initial_high_contrast,
@@ -314,329 +308,3 @@ class AdministratorWindow(Tk):
 
     def get_palm_label(self):
         return self.palm_label
-
-# An instance of this class represents a tab.
-class Page(Frame):
-    def __init__(self, name, window, cap, 
-                on_ear_change, initial_ear, 
-                on_low_contrast, initial_low_contrast,
-                on_high_contrast, initial_high_contrast, 
-                on_min_time_inc, initial_min_time_inc,
-                on_max_time_inc, initial_max_time_inc, 
-                on_new_command, elements, settings_manager, *args, **kwargs):
-
-        self.event_map = {
-            "on_ear_change": on_ear_change,
-            "on_low_contrast": on_low_contrast,
-            "on_high_contrast": on_high_contrast,
-            "on_min_time_inc": on_min_time_inc,
-            "on_max_time_inc": on_max_time_inc,
-            "on_new_command": on_new_command
-        }
-
-        self.initial_value_map = {
-            "on_ear_change": initial_ear,
-            "on_low_contrast": initial_low_contrast,
-            "on_high_contrast": initial_high_contrast,
-            "on_min_time_inc": initial_min_time_inc,
-            "on_max_time_inc": initial_max_time_inc
-        }
-
-        Frame.__init__(self, *args, **kwargs)
-        self.is_debug = False
-        self.is_fps = False
-        self.is_blink_label = False
-        self.is_fist_label = False
-        self.is_palm_label = False
-        self.is_command_field = False
-        self.is_command_menu = False
-        self.has_list_box = False
-        self.gesture_detected = None
-        self.current_command_map = {}
-
-        self.command_manager = settings_manager
-        self.option = 1
-        self.command_index = 0
-        self.is_full = 0
-
-        self.command_links = {}
-        self.new_command = {}
-
-        self.row_index = 1
-        for element in elements:
-            if element["format"] == "text":
-                column_index = 0
-                body_index = list(element.keys()).index("body")
-                for body in list(element.keys())[body_index:]:
-                    if self.is_command_menu:
-                        self.label = Label(self.command_listbox, element.get(body))
-                        self.label.grid(row=self.row_index, column=column_index, padx=10, pady=10)
-                        self.name = name
-                        column_index += 1
-                    elif self.has_list_box:
-                        self.label = Label(self.list_box, element.get(body))
-                        self.label.grid(row=self.row_index, column=column_index, padx=10, pady=0)
-                        self.name = name
-                        column_index += 1
-                    else:
-                        self.label = Label(self, element.get(body))
-                        self.label.grid(row=self.row_index, column=column_index, padx=10, pady=10)
-                        self.name = name
-                        column_index += 1
-
-            elif element["format"] == "logo":
-                self.canvas = Canvas(self.list_box, width=100, height=66)
-                self.canvas.grid(row=self.row_index, column=0, pady=10)
-                image = PIL.ImageTk.PhotoImage(PIL.Image.open("./static_data/NVSHRLogo.png"))
-                self.canvas.create_image(100, 66, anchor="nw", image=image)
-
-            elif element["format"] == "text-cam-status":
-                text_var = StringVar()
-                self.label = Label(self, textvariable=text_var, font=20)
-                text_var.set("Camera On: " + str(cap.isOpened()))
-                self.label.grid(row=self.row_index, column=0, padx=10, pady=10)
-                self.name = name
-
-            elif element["format"] == "text-cam-fps":
-                self.fps_container = StringVar()
-                self.label = Label(self, textvariable=self.fps_container, font=20)
-                self.fps_container.set("FPS:       " + self.get_cam_fps(cap))
-                self.label.grid(row=self.row_index, column=0, padx=10, pady=10)
-                self.name = name
-                self.is_fps = True
-
-            elif element["format"] == "video":
-                self.is_debug = True
-                self.debug_width = cap.get(cv2.CAP_PROP_FRAME_WIDTH)
-                self.debug_height = cap.get(cv2.CAP_PROP_FRAME_HEIGHT)
-                self.debug_canvas = Canvas(self, width=self.debug_width, height=self.debug_height)
-
-                self.debug_canvas.grid(row=self.row_index, column=0, padx=10, pady=10, columnspan=5)
-                self.name = name
-
-            elif element["format"] == "slider":
-                column_index = 0
-                for event in element["events"]:
-                    event_name = event
-                    self.slider_command = self.event_map[event_name]
-                    if event_name == "on_ear_change":
-                        self.slider = Scale(self, orient='horizontal', from_=0, to=10, command=self.slider_command)
-                    elif event_name == "on_low_contrast" or event_name == "on_high_contrast":
-                        self.slider = Scale(self, orient='horizontal', from_=0, to=255, command=self.slider_command)
-                    elif event_name == "on_min_time_inc" or "on_max_time_inc":
-                        self.slider = Scale(self, orient='horizontal', from_=0, to=15, command=self.slider_command)
-
-                    self.slider.set(self.initial_value_map[event_name])  # initial_ear * 100
-                    self.slider.grid(row=self.row_index, column=column_index, padx=10, pady=10)
-                    self.name = name
-                    column_index += 1
-
-            elif element["format"] == "gestures":
-                self.gesturename = Label(self, text=element["body"][0], font=20, bg="White")
-                self.gesturename.grid(row=self.row_index, column=0, padx=10, pady=10)
-
-                self.is_blink_label = True
-                self.blink_label = Label(self, text=element["body"][1], font=20, fg="Blue")
-                self.blink_label.grid(row=self.row_index, column=1, padx=10, pady=10)
-
-                self.is_fist_label = True
-                self.fist_label = Label(self, text=element["body"][2], font=20, fg="Blue")
-                self.fist_label.grid(row=self.row_index, column=2, padx=10, pady=10)
-
-                self.is_palm_label = True
-                self.palm_label = Label(self, text=element["body"][3], font=20, fg="Blue")
-                self.palm_label.grid(row=self.row_index, column=3, padx=10, pady=10)
-
-            elif element["format"] == "button":
-                self.log_button = Button(self, text = 'Click to see contents of the logfile', command = self.open_log_file).pack()
-                self.delete_log_file()
-
-            elif element["format"] == "commands":
-                self.is_command_menu = True
- 
-            elif element["format"] == "listbox":
-                if self.is_command_menu:
-                    self.command_listbox = Listbox(self, bd=0, height=60, width=150)
-                    self.command_listbox.grid(row=self.row_index, column=0, pady=10)
-                else:
-                    self.list_box = Listbox(self, bd=0, height=60, width=150)
-                    self.list_box.grid(row=self.row_index, column=0, pady=10)
-                    self.has_list_box = True
-
-            elif element["format"] == "field":
-                self.field = Entry(self.command_listbox)
-                self.field.grid(row = self.row_index, column = 0)
-                self.is_command_field = True
-
-            elif element["format"] == "new":
-                small_frame = LabelFrame(self.command_listbox, width=1000, height=100, bd=0)
-                small_frame.grid(row=self.row_index, column=0, padx=10, pady=10)
-                for x in range(1, 4):
-                    self.gesture_list = ["fist", "palm", "blink"]
-                    variable = StringVar()
-                    variable.set(" ")
-                    self.new_command[self.command_index] = variable
-                    gesture = OptionMenu(small_frame, variable, *self.gesture_list, command=self.is_full_command)
-                    gesture.grid(row=self.row_index, column=self.command_index, pady=10)
-                    self.command_index += 1
-                add_button = Button(small_frame, text="Add New Command", command=self.add_new_command)
-                add_button.grid(row=self.row_index, column=self.command_index + 1, pady=10)
-
-            elif element["format"] == "device_selection":
-                self.option_list = ["Alexa", "Smart Plug"]
-                row = self.row_index
-                 
-                for option in self.command_manager.action:
-                    small_frame = LabelFrame(self.command_listbox, width=1000, height=100, bd=0)
-                    small_frame.grid(row=row, column=0, padx=10, pady=10)
-                    text = {"text": "Command " + str(self.option)}
-                    self.label = Label(small_frame, text)
-                    self.label.grid(row=row, column=0, padx=10, pady=10)
-                    self.name = name
-                    variable = StringVar()
-                    variable.set(self.command_manager.action[self.option])
-                    self.command_links[self.option] = variable
-                    self.optionMenu = OptionMenu(small_frame, variable, *self.option_list, command=self.set_value)
-                    self.optionMenu.grid(row=row, column=1, padx=10, pady=10, columnspan=100)
-                    self.optionMenu.config(width=30)
-                    self.option += 1
-                    row += 1
-                self.row_index = row
-
-            elif element["format"] == "submit_button":
-                self.log_button = Button(self, text = 'Add Command', command = self.add_command).pack()
-                self.delete_log_file()
-
-            elif element["format"] == "option":
-                self.option_list = ["Alexa", "Smart Plug"]
-                row = self.row_index
-                 
-                for option in self.command_manager.action:
-                    small_frame = LabelFrame(self.command_listbox, width=1000, height=100, bd=0)
-                    small_frame.grid(row=row, column=0, padx=10, pady=10)
-                    text = {"text": "Command " + str(self.option)}
-                    self.label = Label(small_frame, text)
-                    self.label.grid(row=row, column=0, padx=10, pady=10)
-                    self.name = name
-                    variable = StringVar()
-                    variable.set(self.command_manager.action[self.option])
-                    self.command_links[self.option] = variable
-                    self.optionMenu = OptionMenu(small_frame, variable, *self.option_list, command=self.set_value)
-                    self.optionMenu.grid(row=row, column=1, padx=10, pady=10, columnspan=100)
-                    self.optionMenu.config(width=30)
-                    self.option += 1
-                    row += 1
-                self.row_index = row
-
-            self.row_index += 1
-
-    def __bgr_to_rgb__(self, frame):
-        return frame[..., [2, 1, 0]]
-
-    def set_value(self, value):
-        self.current_command_map["device_name"] = value
-        self
-        for x in range(1, 5):
-            if value == self.command_manager.action[x] and value != "None":
-                messagebox.showerror("Smart Home Device Linked", "This smart home device has already been linked to "
-                                                                 "another command. Please chose a different device "
-                                                                 "for this command.")
-                return
-        for x in range(1, 5):
-            self.command_manager.write_to_file(x, self.command_links[x].get())
-
-        self.event_map["on_new_command"](
-                self.current_command_map['gesture_sequence'],
-                self.current_command_map['command_text'],
-                self.current_command_map['device_name']
-                )
-
-    def is_full_command(self, value):
-        self.is_full += 1
-
-    def add_new_command(self):
-
-        if self.is_full >= 3:
-            if self.new_command[0].get() != self.new_command[1].get() and self.new_command[1].get() != self.new_command[
-                2].get():
-                small_frame = LabelFrame(self.command_listbox, width=1000, height=100, bd=0)
-                small_frame.grid(row=self.row_index, column=0, padx=10, pady=10)
-                text = {"text": "Command " + str(self.option) + " (" + self.new_command[0].get() + ", " +
-                                self.new_command[1].get() + ", " + self.new_command[2].get() + ")"}
-
-                self.current_command_map["gesture_sequence"] = [
-                        self.new_command[0].get(), 
-                        self.new_command[1].get(), 
-                        self.new_command[2].get()
-                        ]
-                self.current_command_map["command_text"] = self.field.get()
-                self.label = Label(small_frame, text)
-                self.label.grid(row=self.row_index, column=0, padx=10, pady=10)
-                variable = StringVar()
-                variable.set("None")
-                self.command_links[self.option] = variable
-                self.optionMenu = OptionMenu(small_frame, variable, *self.option_list, command=self.set_value)
-                self.optionMenu.grid(row=self.row_index, column=1, padx=10, pady=10, columnspan=100)
-                self.optionMenu.config(width=30)
-                self.option += 1
-                self.row_index += 1
-            else:
-                messagebox.showerror("Incorrect Command",
-                                     "Commands may not have identical gestures in a successive order.")
-        else:
-            messagebox.showerror("No Command Created", "Please chose three gestures to create a full command.")
-
-        self.is_full = 0
-
-    def open_log_file(self):
-        pdf = FPDF()
-        pdf.add_page()
-        pdf.set_font("Arial", size = 10)
-        count =1
-        file = open('logfile.txt')
-        for line in file:
-            pdf.cell(200, 10, txt = line, ln = count, align = "Left")
-            count +=1
-        file.close()
-        pdf.output("logfile.pdf")
-        subprocess.Popen(["logfile.pdf"], shell = True)
-        
-    def delete_log_file(self):
-        if os.path.exists("logfile.pdf"):
-            os.remove("logfile.pdf")
-  
-    def set_fps(self, fps):
-        self.fps_container.set("FPS:       " + str(fps))
-
-    def __frame_to_image__(self, frame):
-        return PIL.ImageTk.PhotoImage(image=PIL.Image.fromarray(frame))
-
-    def __display_image__(self, image):
-        self.image = image
-        self.debug_canvas.create_image(0, 0, image=self.image, anchor=tkinter.NW)
-
-    def set_debug_frame(self, frame):
-        self.__display_image__(self.__frame_to_image__(self.__bgr_to_rgb__(frame)))
-
-    def get_cam_fps(self, cap):
-        while cap.isOpened():
-            return str(cap.get(cv2.CAP_PROP_FPS))
-
-    # This code, as written, cannot display two simultaneous gestures.
-    def set_gesture_background(self, gesture_detected):
-        if gesture_detected == "fist":
-            self.fist_label.configure(bg="Black")
-            self.palm_label.configure(bg="White")
-            self.blink_label.configure(bg="White")
-        elif gesture_detected == "palm":
-            self.palm_label.configure(bg="Black")
-            self.fist_label.configure(bg="White")
-            self.blink_label.configure(bg="White")
-        elif gesture_detected == "blink":
-            self.blink_label.configure(bg="Black")
-            self.fist_label.configure(bg="White")
-            self.palm_label.configure(bg="White")
-        else:
-            self.fist_label.configure(bg="White")
-            self.palm_label.configure(bg="White")
-            self.blink_label.configure(bg="White")
